restructured ft_printf

+++ milestone01/ft_printf/include/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: ilaliev <ilaliev@student.42heilbronn.de>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/18 15:39:23 by ilaliev           #+#    #+#             */
-/*   Updated: 2025/03/26 13:10:31 by ilaliev          ###   ########.fr       */
+/*   Updated: 2025/07/18 15:45:07 by ilaliev          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,6 @@
 # define FT_PRINTF_H
 
 // librarys
-# include "libft/libft.h"
 # include <stdlib.h>
 # include <unistd.h>
 # include <stdarg.h>
@@ -26,14 +25,14 @@
 int		ft_printf(const char *format, ...);
 int		put_all(char *format, va_list args);
 int		what_case(char format, va_list args);
-int		put_hex(uintptr_t x, char format);
-int		hex_len(uintptr_t p);
+int		put_hex(unsigned long x, char format);
+int		hex_len(unsigned long p);
 char	*ft_uitoa(unsigned int n);
 int		unsigned_num_len(unsigned int n);
 int		print_hex(unsigned int x, char format);
 int		print_sint(int num);
 int		print_uint(int num);
-int		print_adr(uintptr_t p);
+int		print_adr(unsigned long p);
 int		print_str(char *str);
 int		print_c(int c);
